+++ app/components/provider_interface/reference_with_feedback_component.html.erb
@@ -1,0 +1,3 @@
+<div data-qa="reference">
+  <%= render SummaryCardComponent, rows: rows %>
+</div>
