--- app/components/reference_with_feedback_component.html.erb
@@ -1,19 +1,0 @@
-<div data-qa="reference">
-  <%= render SummaryCardComponent, rows: rows do %>
-    <% if @show_send_email %>
-      <%= render(SummaryCardHeaderComponent, title: @title) do %>
-          <div class='app-summary-card__actions'>
-            <% if FeatureFlag.active?('send_reference_email_via_support') %>
-              <%= link_to support_interface_send_reference_email_path(@reference), class: 'govuk-link' do %>
-                Send email<span class="govuk-visually-hidden"> for <%= @reference.name %></span>
-              <% end %>
-            <% else %>
-              <%= link_to support_interface_chase_reference_path(@reference), class: 'govuk-link' do %>
-                Chase referee<span class="govuk-visually-hidden">, <%= @reference.name %></span>
-              <% end %>
-            <% end %>
-          </div>
-      <% end %>
-    <% end %>
-  <% end %>
-</div>
