+++ app/views/support_interface/application_forms/show.html.erb
@@ -56,7 +56,7 @@
   <h2 class="govuk-heading-l govuk-!-margin-top-8">References</h2>
 
   <% @application_form.application_references.each_with_index do |reference, i| %>
-    <%= render ReferenceWithFeedbackComponent, reference: reference, title: "#{(i + 1).ordinalize} reference", show_send_email: true %>
+    <%= render SupportInterface::ReferenceWithFeedbackComponent, reference: reference, title: "#{(i + 1).ordinalize} reference" %>
   <% end %>
 <% end %>
